+++ src/Concerto/PanelBundle/Resources/starter_content/src/test/_CAT/nodes/show_item_6/ports/code.r
@@ -6,14 +6,24 @@ getTemplateParams = function(itemSafe, templateParams, responseRequired, respons
   return(templateParams)
 }
 
-getTimeLimit = function(testTimeStarted, testTimeLimit, itemTimeLimit) {
+getTestTimeLeft = function(testTimeLimitType, testTimeStarted, testTimeLimit, totalTimeTaken, oneSecRound=T) {
   limit = 0
-  startedAgo = as.numeric(Sys.time()) - testTimeStarted
-  testTimeLeft = 0
   if(testTimeLimit > 0) {
-    testTimeLeft = max(testTimeLimit - startedAgo, 1)
-    limit = testTimeLeft
+    if(testTimeLimitType == "startedAgo") {
+      startedAgo = as.numeric(Sys.time()) - testTimeStarted
+      limit = testTimeLimit - startedAgo
+    } else {
+      limit = testTimeLimit - totalTimeTaken
+    }
+    if(oneSecRound) {
+      limit = max(limit, 1)
+    }
   }
+  return(limit)
+}
+
+getTemplateTimeLimit = function(testTimeLimit, itemTimeLimit) {
+  limit = testTimeLimit
   if(itemTimeLimit > 0) {
     if(limit > 0) {
       limit = min(limit, itemTimeLimit)
@@ -36,16 +46,34 @@ if(!concerto.template.isResponseQueued()) {
   }
 }
 
-testTimeLeft = getTimeLimit(
+testTimeLeft = getTestTimeLeft(
+  settings$testTimeLimitType,
   testTimeStarted, 
   as.numeric(settings$testTimeLimit) + as.numeric(settings$testTimeLimitOffset), 
+  totalTimeTaken,
+  T
+)
+
+templateTimeLimit = getTemplateTimeLimit(
+  testTimeLeft,
   as.numeric(settings$itemTimeLimit)
 )
 
 response = concerto.template.show(
   settings$itemTemplate, 
   params=params, 
-  timeLimit=testTimeLeft
+  timeLimit=templateTimeLimit
+)
+
+timeTaken = as.numeric(response$timeTaken)
+totalTimeTaken = totalTimeTaken + timeTaken
+
+testTimeLeft = getTestTimeLeft(
+  settings$testTimeLimitType,
+  testTimeStarted, 
+  as.numeric(settings$testTimeLimit) + as.numeric(settings$testTimeLimitOffset), 
+  totalTimeTaken,
+  F
 )
 
 .branch = "submitted"